+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @appliedblockchain/silentdatarollup-core
 
+## 1.0.11
+
+### Patch Changes
+
+- Hardhat v3 Package added
+- Other improvements and fixes
+
 ## 1.0.10
 
 ### Patch Changes